% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\docType{data}
\name{gene_expr_human}
\alias{gene_expr_human}
\title{The test human gene expression data}
\format{The test human gene expression data contains read counts info about 
152 genes in 12 conditions.}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\usage{
get("gene_expr_human")
}
\description{
HGNC gene symbols in rows, gene expression counts in columns.
}
\keyword{datasets}
