% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenySuppFunc.r
\name{getFullModel}
\alias{getFullModel}
\title{This function is designed for getting a full model matrix}
\usage{
getFullModel(organism)
}
\arguments{
\item{organism}{Model organism has taken from the main function's argument}
}
\description{
This function is designed for getting a full model matrix
}
