% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenySuppFunc.r
\name{getModel}
\alias{getModel}
\title{This function is designed for getting a model matrix with top significant
genes for each pathway}
\usage{
getModel(organism, top)
}
\arguments{
\item{organism}{"Human" or "Mouse" taken from the main function's argument}

\item{top}{Desired top number of genes for each pathway according to their
significance(p.value)}
}
\value{
This function returns model matrix according to the top n significant
}
\description{
This function is designed for getting a model matrix with top significant
genes for each pathway
}
\examples{
#getting a model matrix according to the desired top n significant 
model <- getModel("Human", top=100)
}
