% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{human_perm_expected}
\alias{human_perm_expected}
\title{The test human result output matrix produced by permutation progeny function}
\format{The test human result output matrix contains pathway score for each
condtioin}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\description{
Conditions in rows, pathways in columns.
}
\keyword{datasets}
