% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{input_mouse}
\alias{input_mouse}
\title{The test mouse gene expression data}
\format{The test mouse gene expression data contains read counts info about 
88 genes in 8 conditions.}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\description{
HGNC gene symbols in rows, gene expression counts in columns.
}
\keyword{datasets}
