% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\docType{data}
\name{model}
\alias{model}
\title{The linear model underlying PROGENy}
\format{An object of class \code{matrix} with 1059 rows and 11 columns.}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\usage{
model
}
\description{
HGNC gene symbols in rows, pathways in columns. Pathway activity inference
works by a matrix multiplication of gene expression with the model.
}
\keyword{datasets}

