% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{mouse_def_expected}
\alias{mouse_def_expected}
\title{The test mouse result output matrix produced by default progeny function}
\format{The test mouse result output matrix contains pathway score for each
condtioin}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\description{
Conditions in rows, pathways in columns.
}
\keyword{datasets}
