\name{progeny-package}
\alias{progeny-package}
\alias{progeny}
\docType{package}
\title{
\packageTitle{progeny}
}
\description{
\packageDescription{progeny}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{progeny}
\packageIndices{progeny}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{progeny}

Maintainer: \packageMaintainer{progeny}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
