% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenyPermutations.r
\name{progenyPerm}
\alias{progenyPerm}
\title{\code{progeny_perm}
This function is designed to compute progeny pathway scores and assess there 
significance using a gene sampling based permutation strategy, for a series of
experimental samples/contrasts.}
\usage{
progenyPerm(
  df,
  weight_matrix,
  k = 10000,
  z_scores = TRUE,
  get_nulldist = FALSE
)
}
\arguments{
\item{df}{A data.frame of n*m+1 dimension, where n is the number of omic
features to be considered and m is the number of samples/contrasts.
The first column should be the identifiers of the omic features. 
These identifiers must be coherent with the identifers of the weight matrix.}

\item{weight_matrix}{A progeny coeficient matrix. the first column should be
the identifiers of the omic features, and should be coherent with
the identifiers provided in df.}

\item{k}{The number of permutations to be preformed to generate 
the null-distribution used to estimate significance of progeny scores. 
Default value is 10000.}

\item{z_scores}{if true, the z-scores will be returned for 
the pathway activity estimations. Else, the function returns 
a normalised z-score value between -1 and 1.}

\item{get_nulldist}{if true, the null score distribution used for normalisation
will be returned along with the actual normalised score data frame.}
}
\value{
This function returns a list of two elements. The first element is 
a dataframe of p*m+1 dimensions, where p is the number of progeny pathways, 
and m is the number of samples/contrasts.
Each cell represent the significance of a progeny pathway score for 
one sample/contrast. The signifcance ranges between -1 and 1. 
The significance is equal to x*2-1, x being the quantile of the progeny
pathway score with respect to the null distribution.
Thus, this significance can be interpreted as the equivalent of 1-p.value
(two sided test over an empirical distribution) with the sign indicating
the direction of the regulation.
The second element is the null distribution list (a null distribution is
generated for each sample/contrast).
}
\description{
\code{progeny_perm}
This function is designed to compute progeny pathway scores and assess there 
significance using a gene sampling based permutation strategy, for a series of
experimental samples/contrasts.
}
