% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny-scatter.R
\name{progenyScatter}
\alias{progenyScatter}
\title{This function generate a series of scatter plot with marginal distribution (in the form of an arrangeGrob object), for each progeny pathway and sample/contrast.
Each scatter plot has progeny weights as x axis and the gene level stat used to compute progeny score as y axis.
The marginal distribution of the gene level stats is displayed on the right of the plot to give a visual support of the significnce of each gene contributing to the progeny pathway score.
The colors green and red represent respectivelly positive and negative contribution of genes to the progeny pathway.
For each gene contribution, 4 cases are possible, as the combinaisons of the sign of the gene level stat and the sign of the gene level weigth.
Positive weight will lead to a positive(green)/negative(red) gene contribution if the gene level stat is positive/negative.
Negative weigth will lead to a negative(red)/positive(green) gene contribution if the gene level stat is positive/negative.}
\usage{
progenyScatter(
  df,
  weight_matrix,
  dfID = 1,
  weightID = 1,
  statName = "gene stats"
)
}
\arguments{
\item{df}{a n*m data frame, where n is the number of omic features (genes). m isn't really important, as long as at least one column corespond to a sample or contrast statistic. One of the column should correspond to the gene symboles.}

\item{dfID}{an integer corresponding to the column number of the gene identifiers of df.}

\item{weightID}{an integer corresponding to the column number of the gene identifiers of the weight matrix.}

\item{cm}{a progeny coeficient matrix. One of the column should be the gene symboles.}

\item{statname}{The neame of the stat used, to be displayed on the plot}
}
\value{
The function returns a list of list of arrangeGrob object.The first level list elements correspond to samples/contrasts. The second level correspond to pathways.
The plots can be saved in a pdf format using the saveProgenyPlots function.
}
\description{
This function generate a series of scatter plot with marginal distribution (in the form of an arrangeGrob object), for each progeny pathway and sample/contrast.
Each scatter plot has progeny weights as x axis and the gene level stat used to compute progeny score as y axis.
The marginal distribution of the gene level stats is displayed on the right of the plot to give a visual support of the significnce of each gene contributing to the progeny pathway score.
The colors green and red represent respectivelly positive and negative contribution of genes to the progeny pathway.
For each gene contribution, 4 cases are possible, as the combinaisons of the sign of the gene level stat and the sign of the gene level weigth.
Positive weight will lead to a positive(green)/negative(red) gene contribution if the gene level stat is positive/negative.
Negative weigth will lead to a negative(red)/positive(green) gene contribution if the gene level stat is positive/negative.
}
