% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.r
\docType{data}
\name{result_human_expected}
\alias{result_human_expected}
\title{The test human result output matrix}
\format{The test human result output matrix contains pathway score for each
condtioin}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\usage{
get("result_human_expected")
}
\description{
Conditions in rows, pathways in columns.
}
\keyword{datasets}
