% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{vignette_data}
\alias{vignette_data}
\title{The RNA data used in the progeny vignette}
\format{
List with three elements: the gene counts, the experimental design and 
the result of limma differential analysis        
\describe{
    \item{counts}{gene counts}
    \item{design}{experiemental design}
    \item{limma_ttop}{differential analysis result using limma}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE119931}
}
\description{
List with three elements: the gene counts, the experimental design and 
the result of limma differential analysis
}
\examples{
data("vignette_data")
}
\keyword{datasets}
