% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenySuppFunc.r
\name{getModel}
\alias{getModel}
\title{Returns the Progeny Model}
\usage{
getModel(organism = "Human", top = 100)
}
\arguments{
\item{organism}{"Human" or "Mouse" taken from the main function's argument. 
Default to "Human"}

\item{top}{Desired top number of genes for each pathway according to their
significance(p.value). Default to 100}
}
\value{
This function returns model matrix according to the top n significant
}
\description{
This function is designed for getting a model matrix with top significant
genes for each pathway
}
\examples{
#getting a model matrix according to the desired top n significant 
model <- getModel("Human", top=100)
}
