% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenySuppFunc.r
\name{getModel}
\alias{getModel}
\title{This function is designed for getting a model matrix with top desired number
of genes for each pathway}
\usage{
getModel(full_model, top)
}
\arguments{
\item{top}{Desire top number of genes for each pathway according to their
significance(p.value)}

\item{full}{The full model matrix taken from getFullModel function return}
}
\description{
This function is designed for getting a model matrix with top desired number
of genes for each pathway
}
\examples{
getModel(full_model, top=100)
}
