% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenySuppFunc.r
\name{getModel}
\alias{getModel}
\title{This function is designed for getting a model matrix with top desired number
of genes for each pathway}
\usage{
getModel(full_model, top)
}
\arguments{
\item{full_model}{The full model matrix taken from getFullModel function return}

\item{top}{Desire top number of genes for each pathway according to their
significance(p.value)}
}
\description{
This function is designed for getting a model matrix with top desired number
of genes for each pathway
}
\examples{
#getting a full model matrix
full_model <- getFullModel("Human")

#getting a model matrix according desired top n significant genes
model <- getModel(full_model, top=100)
}
