% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{model_human_full}
\alias{model_human_full}
\title{The full human linear model underlying PROGENy}
\format{The full human model contains 22479 genes, associated pathways,
p.value and p.adj value}
\source{
\url{http://biorxiv.org/content/early/2016/08/28/065672}
}
\description{
HGNC gene symbols in rows, pathways in columns. Pathway activity inference
works by a matrix multiplication of gene expression with the model.
}
\keyword{datasets}
