% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\name{model_mouse_full}
\alias{model_mouse_full}
\title{The full mouse linear model underlying PROGENy}
\format{
The full mouse model contains 17426 genes, associated pathways,
weight and the p-value.         
\describe{
    \item{gene}{gene names in HGNC symbols}
    \item{pathway}{names of PROGENy pathways}
    \item{weight}{z-scores for a given gene}
    \item{p.value}{significance of gene in a pathway}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/31525460}
}
\description{
MGI gene symbols in rows, pathways in columns. Pathway activity inference
works by matrix multiplication of gene expression with the model.
}
\examples{
get("model_mouse_full", envir = .GlobalEnv)
}
\keyword{datasets}
