% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny-package.r, R/progeny.r
\docType{package}
\name{progeny}
\alias{progeny}
\alias{progeny-package}
\title{Calculate PROGENy pathway scores from gene expression}
\usage{
progeny(expr, scale = TRUE)
}
\arguments{
\item{expr}{A gene expression object with HGNC symbols in rows and samples
in columns}

\item{scale}{Logical value indicating whether to scale the scores of each
pathway to have a mean of zero and standard deviation of one}
}
\value{
A matrix with samples in columns and pathways in rows
}
\description{
This function uses the linear model of pathway-responsive genes underlying
the PROGENy method. It transforms a gene expression matrix with HGNC gene
symbols in rows and sample names in columns into a pathway score matrix with
samples and in rows and pathways in columns.

This function uses the linear model of pathway-responsive genes underlying
the PROGENy method. It transforms a gene expression matrix with HGNC gene
symbols in rows and sample names in columns into a pathway score matrix with
samples and in rows and pathways in columns.
}

