% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny-package.r, R/progeny.r
\docType{package}
\name{progeny}
\alias{progeny}
\title{Calculate PROGENy pathway scores from gene expression}
\usage{
progeny(
  expr,
  scale = TRUE,
  organism = "Human",
  top = 100,
  perm = 1,
  ref = rownames(result)
)
}
\arguments{
\item{expr}{A gene expression object with HGNC symbols in rows and samples
in columns. In order to run PROGENy in single-cell RNAseq data,
it also accepts Seurat and SingleCellExperiment object, taking
the normalized counts for the computation.}

\item{scale}{A logical value rnindicating whether to scale the scores of each
pathway to have a mean of zero and a standard deviation of one}

\item{organism}{The model organism - human or mouse}

\item{top}{The top n genes for generating the model matrix according to
significance (p-value)}

\item{perm}{A number of permutations}
}
\value{
A matrix with samples in columns and pathways in rows
}
\description{
This function uses the linear model of pathway-responsive genes underlying
the PROGENy method. It transforms a gene expression matrix with HGNC gene
symbols in rows and sample names in columns into a pathway score matrix with
samples and in rows and pathways in columns.

This function uses the linear model of pathway-responsive genes underlying
the PROGENy method. It transforms a gene expression matrix with HGNC gene
symbols in rows and sample names in columns into a pathway score matrix with
samples and in rows and pathways in columns.
}
\details{
The publication of the method is available at:
https://www.nature.com/articles/s41467-017-02391-6

The supplied expression object has to contain HGNC symbols in rows. This
will, in most cases (and how we originally used it), be either normalized
gene expression of a microarray experiment or log-transformed (and
possible variance-stabilized) counts from an RNA-seq experiment.

The human model matrix itself consists of 14 pathways and 1301 genes 
and the mouse model matrix - 14 and 1376 respectively). Its coefficients
are non-zero if the gene-pathway pair corresponds to the top N genes
(100 by default) that were up-regulated upon stimulation of the pathway 
in a wide range of experiments. The value corresponds to the fitted z-score 
across experiments in our model fit. Only rows with at least one non-zero
coefficient were included, as the rest is not used to infer pathway
activity.
}
\examples{
# use example gene expression matrix here, this is just for illustration
gene_expression <- as.matrix(read.csv(system.file("extdata", 
"human_input.csv", package = "progeny"), row.names = 1))

# calculate pathway activities
pathways <- progeny(gene_expression, scale=TRUE, 
    organism="Human", top = 100, perm = 1)
}
