% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progenySuppFunc.r
\name{saveProgenyPlots}
\alias{saveProgenyPlots}
\title{\code{saveProgenyPlots}}
\usage{
saveProgenyPlots(plots, contrast_names, dirpath)
}
\arguments{
\item{plots}{a list of list of arrangeGrob object (such as the one returned 
by the progenyScatter function.).The first level list elements correspond 
to samples/contrasts. The second level correspond to pathways.
The plots can be saved in a pdf format using the saveProgenyPlots function.}

\item{contrast_names}{a vector of same length as the first level of 
the plot list corresponding to the names of each sample/contrast}

\item{dirpath}{the path to the directory where the plotsshould be saved}
}
\description{
This function is designed to save the plots (in pdf format) of a nested 
(2 level) list of arrangeGrob objects, such as the one returned by 
the progenyScatter function.
}
\examples{
#create plots using progneyScatter function:
plots <- progenyScatter(df, weight_matrix)

#create a list with contrast names:
contrast_names <- names(df[2:ncol(df)])

#assign a path to store your plots
dirpath <- "./progeny_plots/"

#save it
saveProgenyPlots(plots, contast_names, dirpath)
}
